requirements.txt
langchain 
langchain-core 
langchain-openai 
langchain-community 
langchain-experimental
langgraph
langserve[all]
langchain-cli
langsmith
gradio
shutils
pillow
#pytesseract
tesseract-ocr
timestamps
msal
sentence-transformers
matplotlib
openpyxl
pymupdf
chromadb
ollama
langchain-text-splitters
easyocr
extract-msg
python-dateutil